test: separate test into different class
Reduce complexity each test class
